added user logic

+++ demo/src/main/java/com/urlShortener/demo/urlFunctionality/service/UrlServiceImpl.java
@@ -5,12 +5,13 @@ import com.urlShortener.demo.urlFunctionality.entity.UrlAnalytics;
 import com.urlShortener.demo.urlFunctionality.repository.UrlAnalyticsRepository;
 import com.urlShortener.demo.urlFunctionality.repository.UrlRepository;
 import com.urlShortener.demo.userFunctionality.entity.User;
+import com.urlShortener.demo.userFunctionality.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalTime;
-import java.util.Optional;
+import java.util.Collections;
 import java.util.TimeZone;
 
 @Service
@@ -19,16 +20,21 @@ public class UrlServiceImpl implements UrlService {
     private UrlRepository urlRepository;
     @Autowired
     private UrlAnalyticsRepository urlAnalyticsRepository;
+    @Autowired
+    private UserRepository userRepository;
 
 
     @Override
     public String urlCreateService(String longUrl, User loggedInUser) {
-        Url url = new Url(longUrl, loggedInUser);
+        User user = userRepository.findByUserId(loggedInUser.getUserId());
+        Url url = new Url(longUrl, user);
         urlRepository.save(url);
         //change it to look for long link + user
         url = urlRepository.findByLongLinkAndCreatedAt(longUrl, url.getCreatedAt());
-        url.setShortLink(url.encode(url.getId()));
+        url.setShortLink(url.encode(url.getUrlId()));
         urlRepository.save(url);
+        user.addUrl(url);
+        userRepository.save(user);
         return "Url created successfully";
     }
 
@@ -38,23 +44,35 @@ public class UrlServiceImpl implements UrlService {
 
     @Override
     public void deleteAllUrlAnalyticsData(Long id) {
-        urlAnalyticsRepository.deleteAll(urlAnalyticsRepository.findAllByUrl(urlRepository.findById(id).get()));
-    }
-
-    @Override
-    public void updateUrl(Long id, String newLongUrl) {
-        deleteAllUrlAnalyticsData(id);
-        Url url = findById(id);
-        url.setLongLink(newLongUrl);
-        url.setClickCounter(0);
+        Url url = urlRepository.findById(id).get();
+        urlAnalyticsRepository.deleteAllInBatch(url.getUrlAnalytics());
+        url.setUrlAnalytics(null);
         urlRepository.save(url);
     }
 
     @Override
-    public void deleteUrl(Long id) {
+    public String updateUrl(Long id, String newLongUrl, User user) {
         Url url = findById(id);
-        urlAnalyticsRepository.findAllByUrl(Optional.ofNullable(url).get());
-        urlRepository.delete(url);
+
+        if(userRepository.findByEmail(user.getEmail()) != userRepository.findByUrl(url)){
+            return "Invalid User";
+        }
+        deleteAllUrlAnalyticsData(id);
+        url.setLongLink(newLongUrl);
+        url.setClickCounter(0);
+        urlRepository.save(url);
+        return "Url has been created";
+    }
+
+    @Override
+    public String deleteUrl(Long id, User loggedInUser) {
+        Url url = findById(id);
+        if(userRepository.findByUrl(url) != userRepository.findByUserId(loggedInUser.getUserId())){
+            return "Invalid User";
+        }
+        deleteAllUrlAnalyticsData(id);
+        urlRepository.deleteById(url.getUrlId());
+        return "url deleted succesfully";
     }
 
     @Override
@@ -65,8 +83,13 @@ public class UrlServiceImpl implements UrlService {
         }
         Url url = findById(id);
         url.setClickCounter(url.getClickCounter() +1);
-        UrlAnalytics urlAnalytics = new UrlAnalytics(ipAddress, LocalTime.now().toString(), TimeZone.getDefault().getID(), url);
+        UrlAnalytics urlAnalytics = new UrlAnalytics(ipAddress, LocalTime.now().toString(), TimeZone.getDefault().getID());
         urlAnalyticsRepository.save(urlAnalytics);
+        url.addUrlAnalytics(urlAnalytics);
+        urlRepository.save(url);
+        User user = userRepository.findByUserId(url.getUserId());
+        user.addUrl(url);
+        userRepository.save(user);
         return url.getLongLink();
     }
 }
